The file scan-releases.py was added.
This is part of the effort to split pysupport.py into several smaller scripts.
Having several simple scripts seems a better way to implement the functionality
than having a single powerful script with lots of command line options.

scan-releases.py will be the script that is used to scan RELEASE files in an
existing support tree and put the data in a JSON file.