In pys-scan.py group name generation was changed.
The program generates group names from the paths where modules are found.
Before this change it took the last part (basename) of the path to generate the
group name. This sometimes lead to name conflicts since the same name was used
in different subdirectories within the existing support tree. Now the program
uses all parts of the path from a given start point and replaces slashes with
underscore characters when generating the group name.