pys-buildtree.py can now recreate a partial db file.
The db file is the generic dependency database file generated and managed by
pys-db.py. The pys-buildtree.py script can now re-create the db file that was
used to create the buildtree. A partial db file is a file that contains only
parts of a regular db file with exactly one version for each module.

Along with this patch, the user interface of pys-buildtree.py was changed. It
now uses commands in addition to command line options.