In the DB file, dependencies now have a "state".
A state describes how useable and mature a dependency is. Taken from the debian project, we distinguish three states:

* stable  : Tested and working
* testing : Tests are still beeing performed. The buildtree is used by at least
            one application. If this application runs for days or weeks without
            problems caused by the buildtree, we set the dependencies to
            "stable".
* unstable: we try to compile a buildtree with this dependency. If compiling
            works, we change the state to "testing".

In pys-db.py, the command "convert" now requires also a "state", this is the
state that all dependencies get in the generated DB file.

Also on pys-db.py, the command "distribution" now requires a "max state". This
is the maximum allowed state of a dependency in order to be taken into account.
The three possible states have this order:

stable < testing < unstable