pys-buildtree now changes the state of dependencies in the DB file.
If the state of a buildtree is changed, the program also changes the state of
the dependencies in the DB file. There are three possible states, "stable",
"testing" and "unstable". They are defined to have this order:

stable < testing < unstable

If the buildtree is changed to state S, each dependency in the DB file that is
used in the buildtree that has state D is set to min(S,D).

If a new buildtree is created, each dependency it uses that is not present in
the DB file is added with state "unstable". Each dependency it uses that is
present in the DB file with state D is set to min(D,"unstable").