In the DB file, modules now have a state.
Now not just dependencies but also modules have a state. As for dependencies,
we also know three states here, "stable", "testing" and "unstable". They have
the relation:

stable < testing < unstable

If pys-db is used with command "convert", it creates a DB file with all
dependency states and all module states set to the state given on the command
line.

If pys-db is used with command "distribution" and max state M it only takes
modules with state S where S<=M. It also only follows dependencies with state S
where S<=M.

When pys-buildtree is called with command "newtree" it adds modules that are
not already in the DB file with state "unstable".