pys was renamed to sumo.
For now, the project is called SUMO for SUpport MOdule manager.

The old name was discarded since it had "py" for "Python" in it's name but the
project shouldn't get it's name from the language it is implemented in.

This project is an EPICS Support Module Manager, possible abbreviations were:

sm
smm
moma
sum
sumo

From these choices I simply took the one that was more funny than the others.