Support for config files was added in the scripts.
A config file is a JSON file that contains data for some of the command line
options. If an option is not given, the parameter is taken from the config
file.

Each script now has an option "--make-config" that prints the content of a new
config file that is generated from the given command line options to the
console.

Option "--config" is used to specify a config file. If this option is omitted,
the scripts look in the current working directory for a config file that is the
script name appended with ".config.".