sumo-scan now also scans CROSS_COMPILER_TARGET_ARCHS.
This variable defines the target architectures a module is compiled for. It is
often but not always defined in file "config/CONFIG".

Since this macro may be defined elsewhere, the script calls "make" in the
"config" directory if a makefile exists. Otherwise it calls "make" directly
with file "config".

The generated JSON file now has a new key, "archs" where all found target
architectures for all found paths are collected.