"sumo-db convert" now also converts target architectures.
sumo-scan now also collects the target architectures of each module. This new
version of sumo-db stores these with key "archs" for each moduleversion in the
created db file.

Since target architectures are basically a set, but JSON doesn't have direct
support for sets, these are stored as maps mapping a target architecture name
to the value "true". The value "false" is treated as if the architecture is not
supported, although the common way for unsupported architectures would be not
to list them at all.