Further architecture support was added, mostly in sumo-db.
For the target architecture support we need to have at least one target
architecture defined for *every* module. sumo-scan only scans architectures for
all directories below the given directory. This leads to the unwanted situation
that the generated SCAN file can contain references to modules that are in
directories not below the given one with no target architecture defined.

With this patch you can specify more than one directory for sumo-scan, all
directories below the given ones are scanned and the target architectures for
the modules are collected.

sumo-db now requires the new option "--arch" for the commands "distribution",
"shownewest" and "showall".

Some changes had to be applied in sumo-build since some method of class
"Dependencies" in utils.py now *require* an "archs" parameter. Target
architecture support in sumo-build however, is still rudimentary.