Bugfix in sumo-db, cmd "distribution": multiple dependencies.
If a module has multiple dependencies, there are more than one possible
versions of another module it depdends on.

If sumo-db creates a distribution it is necessary that the created partial-db
file has no more than one version of any dependency module.

With this fix, multiple dependencies are removed in the generated partial-db
file.