sumo-scan now scans the supported architectures differently.
Instead of looking for the makefile macro "CROSS_COMPILER_TARGET_ARCHS", the
program now looks for the directories "bin" and "lib". It takes the supported
architectures from the subdirectories of these.

Although this requires that the scanned modules are build, otherwise the
directories don't exist, this is a much more solid method to look for supported
architectures.

Another feature of this is that *all* supported architectures are found, not
only the ones of the cross compiler, e.g. "linux-x86".