builds may now have multiple architectures.
In order to use sumo with an important application here at HZB we had to remove
the convention that all modules of a build and compiled for the same set of
architectures. Now the modules are built for all architectures they support.
Information on the set of supported architectures of a module is only held in
the DB file.