The comparison algorithm of repository tags was changed.
When the user specifies a module in the form "module:version" the program now
looks for an *exact* match of the given version. Here is an example, if the
user specifies:

SEQ:seq-2-3-12

In the old implementation, the program would match this with tag "R-2-3-12" and
"seq-2-3-12", since both contain the numbers "2-3-12".

Now the program matches only "seq-2-3-12", we assume that the user always wants
to match the *exact* string of the version if he doesn't specify a version
range like in "SEQ:-seq-2-3-12". In the latter case the program would still
match "R-2-3-12", "seq-2-3-12" and of course "R-2-3-11", "seq-2-3-11" and so
on.