In sumo-scan, the "--hint" option was added.
This option also replaces "--accept-path". The new "--hint" option is more generic. It's argument has the form "REGEXP,FLAG{,FLAG}". The REGEXP regular expression is matched with the module path. The flag is one of a list of known strings that determines how the support is scanned.

Currently there are two possible FLAG values:

PATH   : treat the support as if there was no repository found, just take the
         path as a source.

TAGLESS: treat the support as if there was no regular repository tag found,
         just take the working copy repository as repository source.