In classes Dependencies and Builddb, a self-check was added.
Both classes uses generic functions to read data from a JSON file. It may be
the case that a file contains valid JSON data but that this data is not a valid
Dependencies or Builddb object. Now both classes do a simple consistency check
after loading a file and raise an exception if the data is not valid.