Weights have been introduced to sumo-db and sumo-build.
A weight factor is an optional integer that can be added to any version of a
module. If sumo-build creates a RELEASE file, that weight factor is the first
order sort key. It is guaranteed that the weight factors of modules in a
RELEASE file are in ascending order. If a weight is not defined for a
moduleversion, a weight factor of 0 is assumed.