In sumo-build command "cleanup" was added.
When "sumo-build new" fails with an exception it can happen that some of the
directories of the build are created while some are missing. In such a
situation you want to clean up your support directory by removing all
directories that belong to the failed build.

Now the program creates a JSON file named "cleanup-[buildtag]". If the
"sumo-build new" exists unexpectedly, this cleanup file can be used to remove
the remains of the failed build.

If "sumo-build new" runs successfully, the cleanup file is removed
automatically.