The test directory was reorganized and improved.
Now we have new names of the test scripts that give a hint on the command that
is tested. So instead of "test07.sh" we have "070-sumo-db-show.sh".

The number at the start determines the order in which the tests are run.

Also, dependencies between the tests are now stored in the test scripts. If the
scripts are called with "deps" and a command line argument, they print the
dependencies of the test in a format that can directly be used by make. This is
used by make to create a depedency file, called "deps" that is then included by
the makefile.