In sumo-scan more options can now be space separated strings.
Some options in sumo-scan can be put several times on the command line in order to give a list of values. Some of these options now also allow to be called with a single string that contains a space separated list of values. So this:

  sumo-scan -d dir1 -d dir2

can also be entered like this:

  sumo-scan -d 'dir1 dir2'

These are the options that allow this:

  -d --dir
  -N --ignore-name
  -g --group-basename
  --exclude-path