In sumo-build command "new" now also calls make.
Command "new" not only creates the source files for the build but also calls
"make". If make was successful, the state of the build is set to "testing".

The new option "--no-make" can prevent the program from calling make.

Option "--makeopts" is used to pass command line options to the make program.