In sumo-build commands "use" and "useauto" were merged.
"useauto" meant that the user doesn't have to provide a buildtag since the
program determines a matching buildtag itself.

Now "use" can also be started without a buildtag so it replaces "useall". In
order to do this, buildtag had to be implemented as a command line option. In
order to keep commands like "state" simple, these allow the specification of
the buildtag as command line option *or* parameter.