The "partialdb" concept was removed.
Before this patch, sumo-db was used to create a partial dependency database
file for a build. sumo-build then had to read this file in order to create the
build. The effect was the two programs and an intermediate file were needed to
create a build from a list of module specifications.

Now sumo-build can create a build from a list of module specifications
directly. The command "distribution" was removed from sumo-db and it's function
was added to the "new" command in sumo-build.