Bugfix in module specifications: user errors were to handled correctly.
Some module specification commands need an argument, e.g. ":rm:module:version".
If the user forgets the argument like in ":rm", the program should print an
error message and exit. In ModuleSpecs._from_strings() this error was not
recognized which lead to an exception.

Now the program prints an error message and exits gracefully.