With sumo-build "new" Option "--add-deps" allows to add missing dependencies.
Usually, when a set of modules is given to "sumo-build new", the program checks
if all these are represented in the dependency database and all dependencies
there are satisfied. If not, the program stops with an error message.

Now, when "sumo-build" is started with option "--add-deps", the program may add
dependencies to the dependency database in order to make it complient with the
given set of versioned modules.