Bugfix: JSON files had trailing spaces on newer python versions.
This is a bug in the JSON module of the standard python library. It adds spaces
at lines that end with a comma ",".

This bug seems only to be fixed in python 3.4 or newer. Now some extra code
does a "rstrip()" on each line in generated JSON files.

Many tests were changed that previously had extra code to handle this bug.