Bugfix in the module search algorithm (sumo-build commands "try" and "new").
For commands "new" and "try" the program gets a module specification, a list of
unversioned and versioned modules, and creates a complete set of versioned
modules from there.

This set is initially empty and is filled by the algorithm with versioned
modules. The program takes the first module of the module specification, adds
all direct and indirect dependencies to the set, and continues with the next
module specification.

Bofore this patch, the program used what looks like the newest version, for
each dependency. This causes problems if somewhere in the module specifications
that follow another version of that dependency is required.

Now the program always looks first at all module specifications if it needs to
determine which version of a dependency it should use.