The format of a dependency database was changed, no more dependency versions.
The dependencies of a module are now just a list of module names of
dependencies. The database no longer contains version names of dependencies.

The reason for this is that the builds in the build database give a better
picture of which version of a dependency can be built with a version of a
module. As a first step for this, class Dependencies in sumo/Databases.py was
changed.

Along with this change, the field "state" in the dependency database was also
removed. How usable a version should also be determined by examining the build
database.