Support for creating and using the scan database file was added.
This file basically contains the information that was lost when the format of
the dependency database was changed. It contains for each dependency the
version of the dependency module as well as a "state". The state shows how
reliable it is to use the module and the dependency together in a build.

The datastructure is interally called a Buildcache since it also has the
purpose to hold additional information retrieved from the build database.

The Buildcache is created by "sumo-db convert-old" and "sumo-db convert".

"sumo-build try" uses the Buildcache to show what module versions where used
with what dependency versions. If "--scandb" is given as a command line
argument, the Buildcache is initialized with that file before data from the
build database is added, otherwise it just uses the build database.