Configuration files can now be merged with command line options.
This is a change for "list options". These are command line options that can be
given more than once on the command line in order to specify a *list* of items.
If such a list option is specified in the configuration file and the command
line, values from the configuration file are replaced by the command line.

Now, if the user uses --mergeoption <optionname>, the two lists for this
options are actually merged, meaning that the result is a combination of both
lists in which each element is unique.

This gives more flexibility in using configuration files and in modifying
configuration files on the command line if the "makeconfig" command is used.