Command "makeconfig" can now save subsets of command line options.
The command now must always be followed by a filename. After this filename, the
user may specify a list of optionnames. If optionnames are specified, only
these are saved in the new created configuration file.

This allowes to change and extend for example "modules" files, that are
basically configuration files with just the two optionnames "alias" and
"module".