In sumo-db command "cloneversion" was improved.
The user may now use this command without a source specification, in this case
the version name is added as a tag to the source specification of the module
that is cloned. The program now always shows what it would changed and asks the
user for a confirmation unless option "-y" is used.