The file locking was rewritten.
Sumo no longer depends on the python module "lockfile". It now implements a
simple file locking itself.

On the linux platform it creates a symbolic link for each lock, on other
platforms it creates a directory.

The file locking mechanism is quite general with the exception that in the
directory where a file is locked there should be no file that has the name
*.lock.

The new symbolic link locking is also supported by the sumo-edit tool, which is
an editor for the dependency database file.