A new state for builds, "disabled", was added.
This is used for builds that may be in use by applications but still shouldn't
be used again. This may be if the parts of the dependency database these builds
used were changed, meaning they couldn't be reproduced.

As a special feature, "sumo state <BUILD> disable" also disables, after
confirmation by the user, all builds that depends on the given build.