Bugfix in makefile_scan.py, environment variables could break the scan.
The problem was that "make" was called with option "-e" which means that
environment variables override settings from the makefile. When there was an
environment variable set that had the same name as a variable in the makefile,
the scan was no longer correct. The bugfix was easy, simply remove "-e" from
the call to "make".