Automatic merging of modulespecs from different sources was removed.
Before this patch, modulespecs from the configuration file, the "-m" command
line option and command arguments were always merged. This lead to some
unexpected behavior. When for example "sumo db find MODULE" was entered, the
command merged MODULE with all modules from sumo.config and showed information
for all these modules. This was certainly not what the user expected.

Now the whole thing was made more generic. As a default, options *override*
settings from the config file. If they are to be concatenated, the user has to
use option "--append" or "-A".

This option is implemented in both scripts, sumo and sumo-scan.

If for example the user wants for "sumo build try" to add a module to the list
of modules from the configuration file, the command would be:

sumo build try MODULE -A module

This is slightly more complicated than "sumo build try MODULE" before this
patch, but other commands like "sumo db find MODULE" behave now like expected.