The "archs" property in the dependency database was removed.
The concept of "archs" for selecting modules was in the end never used. Our
applications are collections of modules that do not have a single set of
architectures. They instead consist of modules that are cross-compiled and
modules that are just compiled for the host architecture.

With this patch not only the "archs" property was removed but all other support
and command line options concerning architectures were removed.