Lists elements in load commands in config files are now unique.
Usually commands like "#opt-preload" in configuration files are executed
immediately. The file is loaded and added to the current configuration.

With sumo command line option "--disable-loading" however, these commands are
treated like any other configuration file entries. If two files contain a load
command like "#opt-preload" the lists of files are simply concatenated. If this
is used to create a new configuration file it may happen that in the resulting
load command files are mentioned twice. With this patch this can no longer
happen, each file can occur in the list only one time.