Bugfix and change in sumo db cloneversion.
sumo db cloneversion issued a warning about an unchanged source specification
when in fact it did change (the tag was removed). This is now fixed. Another
change is that instead of "*" as a wildcard which must be quoted in order not
to interfere with the shell we can now use "." as a wildcard, too.