Bugfix in some uses of system(), stderr messages were not printed.
Function sumolib.system.system() raises an exception, when the command fails.
In this case it *does not* return stdout or stderr even if it was asked to do
so. In case we want to catch errors and *always* catch stderr, we must use
sumolib.system.system_rc(). This function never raises an exception but returns
the command's system return code instead.