The edit command was re-written, it can now only be used to edit DEPS.DB.
Before this patch, the edit command could be used to edit an arbitrary file. In
real life we only want to edit DEPS.DB, so the command is no longer a
maincommand but a sub-command to db. It is invoked with "sumo db edit". The
command now checks if the file is still a valid JSON file and commits the
changes if option --dbrepo was given.

This makes the command really useful for all changes in DEPS.DB that are too
complex in order to be done by command line options. Since DEPS.DB is
re-written by sumo, it is also formatted in a standard way.