In module JSON, regular temporary filenames are used instead of *.tmp.
Function dump_file in this module needs to create a temporary file. Before this
patch the name of this file was simply the filename with a ".tmp" attached. Now
a regular temporary filename is created with tempfile.mkstemp. The need for
this patch arose from a name conflict with another module created filenames the
same way.