Automatic commits can now start an editor for entering a log message.
Before this patch, sumo caught stdout from the commit command, but this made it
impossible for an editor to be run by the version control system. Now sumo only
catches stdout where it's needed and only when there was a log message provided
with option --logmsg. All of this is only relevant when a version controls
system is used to manage the dependency database (option --dbrepo).