Bugfix: BUILDS.DB was changed to have faulty file permissions.
When this file was changed it was recreated with wrong file permissions. The
file was only readable and writeable by it's owner, no one else, not even the
user's group. This had to do with permissions that are used when a temporary
file is created with tempfile.mkstemp(). Now we copy the file permissions from
the original file to the new version of the file with function
shutil.copymode(). This should also work on non unix systems like windows.