Flush stdout and stderr when something is printed on stderr.
When the program prints messages first on stdout and later on stderr it is not
clearly defined in what order the messages will appear on the screen. This is a
problem for tests which rely on a reproducible output.

Now at various places where the program prints messages to stderr it first
flushes stdout, then prints to stderr, then flushes stderr.