Dependencies.DB.selfcheck() now works reproducible.
Before this patch, the key of the object that DB.selfcheck() examined was
arbitrary, it depended on the undefined order of keys in a dictionary. Now the
function uses a sorted list of dictionary keys, meaning that it's results are
now the same for each run of a program and for different versions of python.