Porting to python 3, new subdirs, changes in administration tools.
We want to port sumo to python 3 but we want to keep the python 2 version of
sumo for now. In order to do this, the complete source tree (directories bin
and sumolib) is copied. In directory "python2" there are the python 2 sources,
in directory "python3" there are the python 3 sources. Some scripts in
directory "administration_tools" had to be changed, too.

The makefile in directory "test" now requires that variable "PYVER" is set to 2
or 3.

Note that the sources in directory "python3" are not yet real python 3 sources
but simply copies of the files in directory "python2". The files in "python3"
will be changed in a later patch.