The exception handling in sumo was changed.
Throughout the program all places where exceptions may be raised were marked
with comments. The exceptions are caught only at one place in function process.
If option --exceptions is given, exceptions are not caught but the program
terminates with a stack trace as usual in python. Otherwise the program
terminates more user friendly with a simple error message to stderr.

The purpose of this change is to improve maintainability of the program.