Bugfix in darcs.py, the recognition of local patches was wrong.
When running "darcs push ... --dry-run" with no patches to push, darcs in the
past replied:

  No recorded local changes to push!

In the most recent version of darcs it replies:

  No recorded local patches to push!

The program didn't find the first string in the reply and assumed that there
were patches not yet pushed. This in the end made the test sumo-scan-all fail.
Now the program just looks if the string darcs replies starts with "No recorded
local".