Bugfix in mercurial.py, conflict detection after "pull" was broken.
"hg pull -q -u" does not automatically start merge. Since this is combined with
option "-q" then mercurial does not even show that there is the necessity to
merge.

Now the program does "hg pull -u" and scans for the message "(+1 heads)" with a
regular expression. If this is found, the program starts "hg merge".