Make "sumo build delete" safer when files cannot be deleted.
If "sumo build delete" failed since some files couldn't be deleted, the build
directory was left in an undefined state. Now the program ensures that the
build database "BUILDS.DB" is always in a defined state.

When the program fails in the middle of the command, builds that are
half-deleted remain in the build state "broken". This means that such a build
is unusable and must be deleted manually.