A function to parse NAME=VALUE [,VALUE] was added to sumolib.utils.
The NAME must be a regular identifier like it is allowed in python.

The VALUE may be any simple (non compound) JSON data type. Note that a string
must be enclosed in double quotes if it contains one of the '"', ',' or ' '
characters.

VALUE may be follwed by a comma ',' and another VALUE.

The function sumolib.utils.definition_list_to_dict() takes a string and returns
a dict.

This is a preparation for a more generic source specification on the command
line.