Bugfix in JSON.Container: the __del__ method sometimes failed.
When a derived class of JSON.Container raised an exception before __init__() of
JSON.Container was called, the resulting object would remain in a not properly
constructed state.

But python would call the __del__() method when the object was deleted or
garbage collected anyway. In this case, self.unlock_file() which was called
from __del__() would then fail.  Now __del__() first tests if the object has an
attribute "dict_". Only when this is found, it is assumed that
self.unlock_file() can be called safely.