sumo: Empty make recipes can be defined for modules that have no makefile.
With

  sumo make-recipes MODULE

you define an empty make recipe section for a MODULE. This must be done for
modules that have no makefile at all.

Note that this is slightly incompatible with the previous version of sumo that
probed for a file named "Makefile" or "makefile" and, if it wasn't found,
didn't try to run make for that module.

Now a module without a makefile *must* have an empty make-recipes section
defined in the dependency database.