Abort with an error message if a build does not contain all it's modules.
Usually a build that is not in state "unstable", "incomplete" or "broken"
should contain all modules according to the dependency database. It may
however, happen that this condition is no longer true when the dependency
database was edited a build was created. With this patch sumo aborts in this
case with a proper error message.