The test documentation now shows all steps to bootstrap the tests.
Some extra python libraries have to be created before the tests can be started.
These steps are now documented in test/README.rst.

For this, the script administration_tools/doc-rebuild.sh was splitted. The new
script mk-sumo-doc.sh creates the python libaries for the online documentation
and is called by doc-rebuild.sh.