The exception helper function "annotate" was moved from sumo to utils.py.
This function takes an exception object, prepends a text to the exception text
and returns a new exception object. This is used at some places to add
additional information to exceptions.

A direct call to sys.exit was at some places in sumo replaced by
sumolib.utils.exception_exit(), this function makes a nicer error message.