Bugfix: At python 3.2.3, errors while checking out modules were not recognized.
This has to do with the use of the multiprocessing module that is used for for
checking out support modules. It only happens with quite old python 3 versions,
like 3.2.3.

Here sys.exit(STRING) doesn't work together with multiprocessing.process(). The
main process doesn't get a non-zero error code as expected. As a fix we have to
separate printing to stderr from returning an error code and make them two
separate statements like this:

  sys.stderr.write(STRING)
  sys.exit(1)