sumo --readonly implies repomode "get".
This means that with "--readonly" missing write access to "DEPS.DB" doesn't
cause a warning to be printed even if dbrepomode was not "get". We assume that
"--readonly" means we never *intended" to change "DEPS.DB", so there should be
no warning if the file cannot be written to (and the file lock cannot be
created).